Crate pallet_block_fees

Pallet Domain Transaction Fees

Modules

dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

fees

storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Structs

Pallet

Pallet block-fees to store the accumulated rewards of the current block

Enums

Call

Contains a variant per dispatchable extrinsic that this pallet has.

Traits

Config

Configuration trait of this pallet.

Type Aliases

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.